requirements.txt
crewai==0.40.1
streamlit==1.41.1
mistralai==0.1.11
langchain==0.2.14
langchain-community==0.2.14
langchain-mistralai==0.1.5
python-dotenv==1.0.0
duckduckgo-search==3.9.11
litellm==1.80.8
pydantic==2.12.5
aiohttp==3.12.15
openai==2.9.0
httpx==0.28.1
jinja2==3.1.6
pytz==2024.1
